+++ src/components/CartSection/CartTable/CartTable.tsx
@@ -17,7 +17,7 @@ import { FormButton } from '@/components/LoginForm/FormButton.tsx'
 import { toast } from 'react-hot-toast'
 import type { TCartItem } from '@/types/user-types.ts'
 import { MdDeleteForever } from 'react-icons/md'
-import { calculatePrices, findPromoCodes } from '@/shared/utilities/type-utilities.ts'
+import { calculatePrices, findPromoCodes } from '@/shared/utilities/commerceTools-utilities.ts'
 
 export const CartTable = (): JSX.Element => {
   const { data, error, loading, refetch } = useFetch<ClientResponse<Cart>>(api.cart.fetchActiveCart)
