+++ src/shared/utilities/price.ts
@@ -1,2 +1,2 @@
-export const CENTS_IN_DOLLAR = 100
+export const CENTS_IN_EURO = 100
 export const DECIMAL_PLACES = 2
